Refactor color extraction logic by removing outdated scripts and updating imports. The `updateChannelColors` and `updatePodcastColors` functions have been relocated to the `server` directory for better organization. Additionally, the `extractVibrantColors` function has been moved to the `server` directory, and related imports have been updated throughout the codebase. This change enhances maintainability and aligns with the project's structure.
